feat: resolve recursive reference
