+++ src/index.ts
@@ -1,11 +1,10 @@
-import { UpdatableConfig } from './types/config'
-
 export * from './account'
 export * from './amount'
 export * from './app'
 export * from './app-client'
 export * from './app-deploy'
 export * from './asset'
+export * from './config'
 export * from './debugging'
 export * from './dispenser-client'
 export * from './indexer-lookup'
@@ -13,6 +12,3 @@ export * from './localnet'
 export * from './network-client'
 export * from './transaction'
 export * from './transfer'
-
-/** The AlgoKit config. To update it use the configure method. */
-export const Config = new UpdatableConfig()
